made some changes to structuring of mesh ingress code
